Add basic output check
